+++ zjit/bindgen/src/main.rs
@@ -51,6 +51,7 @@ fn main() {
 
         // Our C file for glue code
         .header(src_root.join(c_file).to_str().unwrap())
+        .header(src_root.join("jit.c").to_str().unwrap())
 
         // Don't want to copy over C comment
         .generate_comments(false)
